requirements.txt
yt-dlp
requests
#torch
#torchaudio
openai-whisper